fix: align confidence_score type to DOUBLE across all definitions
Python float is 64-bit (double), but the audit log table DDL and schema
definitions used FLOAT (32-bit). When write_audit_log created a DataFrame
from pandas, Spark inferred DoubleType which conflicted with the FLOAT
column, causing DELTA_FAILED_TO_MERGE_FIELDS.

Changes:
- 01_setup_tables.py: confidence_score FLOAT → DOUBLE
- delta_tables.py: confidence_score float → double, added DoubleType import
- delta_writer.py: use explicit QC_AUDIT_LOG_SCHEMA when creating DataFrame

+++ notebooks/01_setup_tables.py
@@ -30,7 +30,7 @@ CREATE TABLE IF NOT EXISTS {catalog}.{schema}.qc_audit_log (
     check_type      STRING NOT NULL,
     original_value  STRING,
     corrected_value STRING,
-    confidence_score FLOAT,
+    confidence_score DOUBLE,
     support_level   STRING,
     was_applied     BOOLEAN,
     correction_method STRING,
